+++ app/pico/implementations/lcd_screen.h
@@ -22,12 +22,12 @@ public:
 	void draw();
 
 	void clear() override;
-	void draw_line(uint start_x, uint start_y, uint end_x, uint end_y, uint thickness) override;
+	void draw_line(uint8_t start_x, uint8_t start_y, uint8_t end_x, uint8_t end_y, uint8_t thickness) override;
 	void draw_header(std::string string) override;
-	void draw_divider_horizontal(uint position_px, uint thickness) override;
-	void draw_divider_vertical(uint position_px, uint thickness) override;
-	void draw_text(std::string text, uint x, uint y, FontSize size) override;
-	void draw_circle(uint center_x, uint center_y, uint radius, Color color) override;
+	void draw_divider_horizontal(uint8_t position_px, uint8_t thickness) override;
+	void draw_divider_vertical(uint8_t position_px, uint8_t thickness) override;
+	void draw_text(std::string text, uint8_t x, uint8_t y, FontSize size) override;
+	void draw_circle(uint8_t center_x, uint8_t center_y, uint8_t radius, Color color) override;
 
 private:
 	uint16_t* _display_buffer{};
@@ -43,7 +43,7 @@ private:
 	uint16_t _color_foreground = WHITE;
 	uint16_t _color_background = BLACK;
 
-	static DOT_PIXEL convert_thickness(uint thickness);
+	static DOT_PIXEL convert_thickness(uint8_t thickness);
 	static uint16_t convert_color(Color color);
 	static _tFont* convert_font(FontSize size);
 };
